Arreglado ASEmployeeImp y terminado de implementar la operacion getTotalSalary de MainOffice

+++ src/business/employee/Employee.java
@@ -57,7 +57,7 @@ public abstract class Employee implements Serializable {
         this.active = te.isActive();
     }
 
-    public abstract Double getDetailedSalary();
+    public abstract Float getDetailedSalary();
 
     public static long getSerialVersionUID() {
 		return serialVersionUID;
